chore: rename vendored ag-ui module to ag-ui-schemas.ts
Matches the sibling naming convention in schemas/ (event-schemas.ts,
response-schemas.ts) and reflects that it holds the AG-UI message + event
schemas.
